Move BuildCacheControllerFactory to :core
